Indexing/requirements.txt
Django==3.2.18
django_haystack==3.2.1
pandas==1.5.3